# backend/docs/swagger/swagger.yaml
basePath: /api
definitions:
...
  github_com_MateWorkspace_mate-things_backend_internal_presentation_http_request.AuthLoginRequest:
    properties:
      password:
        example: BrewMeUp!42
        type: string
      username:
        example: grace.hopper
        type: string
    type: object
  github_com_MateWorkspace_mate-things_backend_internal_presentation_http_request.AuthRefreshRequest:
    properties:
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJncmFjZS5ob3BwZXIifQ.brewbrewbrew
        type: string
    type: object
...
  github_com_MateWorkspace_mate-things_backend_internal_presentation_http_request.ProfilePasswordPatchRequest:
    properties:
      current_password:
        example: BrewMeUp!42
        type: string
      new_password:
        example: EvenMoreEspresso!7
        type: string
    type: object
  github_com_MateWorkspace_mate-things_backend_internal_presentation_http_request.ProfilePatchRequest:
    properties:
      bio:
        example: Keeps the espresso machines humming and the firmware fresh.
        type: string
      name:
        example: Grace Hopper
        type: string
      username:
        example: grace.hopper
        type: string
    type: object
...
  github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse:
    properties:
      details:
        example: node_class_id is required
        type: string
      error:
        example: Invalid Format
        type: string
      message:
        example: Please select a valid node class.
        type: string
    type: object
...
  github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.LoginResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJncmFjZS5ob3BwZXIifQ.espresso-shot-pulled
        type: string
      permissions:
        items:
          $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.PermissionResponse'
        type: array
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJncmFjZS5ob3BwZXIifQ.decaf-is-not-an-option
        type: string
      role:
        $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.RoleResponse'
      user:
        $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.UserResponse'
    type: object
...
  github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.PermissionResponse:
    properties:
      created_at:
        example: "2026-06-15T09:30:00Z"
        type: string
      created_by:
        example: e1f4b7c0-2d5e-4f8a-9b3c-6e0f2a5d8c01
        type: string
      deleted_at:
        example: "2026-07-20T11:00:00Z"
        type: string
      deleted_by:
        example: e1f4b7c0-2d5e-4f8a-9b3c-6e0f2a5d8c01
        type: string
      description:
        example: View registered nodes and their status.
        type: string
      id:
        example: 1d2e3f4a-5b6c-4d7e-8f9a-0b1c2d3e4f5a
        type: string
      name:
        example: node:get
        type: string
      preferences:
        type: object
      updated_at:
        example: "2026-07-01T14:05:00Z"
        type: string
      updated_by:
        example: e1f4b7c0-2d5e-4f8a-9b3c-6e0f2a5d8c01
        type: string
    type: object
...
  github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.RoleResponse:
    properties:
      created_at:
        example: "2026-06-15T09:30:00Z"
        type: string
      created_by:
        example: e1f4b7c0-2d5e-4f8a-9b3c-6e0f2a5d8c01
        type: string
      deleted_at:
        example: "2026-07-20T11:00:00Z"
        type: string
      deleted_by:
        example: e1f4b7c0-2d5e-4f8a-9b3c-6e0f2a5d8c01
        type: string
      description:
        example: Can dispatch brew actions on the shop floor, but can't manage the
          fleet.
        type: string
      id:
        example: 7a3c2e10-4b1a-4b8e-9f3a-2b6e7c9d1a04
        type: string
      is_default:
        example: false
        type: boolean
      name:
        example: barista
        type: string
      preferences:
        type: object
      updated_at:
        example: "2026-07-01T14:05:00Z"
        type: string
      updated_by:
        example: e1f4b7c0-2d5e-4f8a-9b3c-6e0f2a5d8c01
        type: string
    type: object
...
  github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.UserResponse:
    properties:
      bio:
        example: Keeps the espresso machines humming and the firmware fresh.
        type: string
      created_at:
        example: "2026-06-15T09:30:00Z"
        type: string
      created_by:
        example: e1f4b7c0-2d5e-4f8a-9b3c-6e0f2a5d8c01
        type: string
      deleted_at:
        example: "2026-07-20T11:00:00Z"
        type: string
      deleted_by:
        example: e1f4b7c0-2d5e-4f8a-9b3c-6e0f2a5d8c01
        type: string
      id:
        example: e1f4b7c0-2d5e-4f8a-9b3c-6e0f2a5d8c01
        type: string
      name:
        example: Grace Hopper
        type: string
      preferences:
        type: object
      role_id:
        example: 7a3c2e10-4b1a-4b8e-9f3a-2b6e7c9d1a04
        type: string
      updated_at:
        example: "2026-07-01T14:05:00Z"
        type: string
      updated_by:
        example: e1f4b7c0-2d5e-4f8a-9b3c-6e0f2a5d8c01
        type: string
      username:
        example: grace.hopper
        type: string
    type: object
info:
  contact: {}
  description: Nusapala Things is the public HTTP and MQTT entrypoint for IoT telemetry,
    action dispatch, and firmware OTA.
  title: Nusapala Things
  version: "1.0"
paths:
...
  /v1/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_request.AuthLoginRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.LoginResponse'
        "400":
          description: Invalid Format
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Auth Login
      tags:
      - Auth
  /v1/auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_request.AuthRefreshRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.LoginResponse'
        "400":
          description: Invalid Format
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "401":
          description: Session Expired / Invalid Token
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Auth Refresh
      tags:
      - Auth
...
  /v1/profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "403":
          description: Access Denied
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Profile Get
      tags:
      - Profile
    patch:
      consumes:
      - application/json
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_request.ProfilePatchRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid Format
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "403":
          description: Access Denied
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "409":
          description: Already Exists
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Profile
      tags:
      - Profile
  /v1/profile/password:
    patch:
      consumes:
      - application/json
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_request.ProfilePasswordPatchRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid Format
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "403":
          description: Access Denied
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Profile Password
      tags:
      - Profile
  /v1/profile/permissions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.PermissionResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "403":
          description: Access Denied
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Profile Permissions Get
      tags:
      - Profile
...
securityDefinitions:
  BearerAuth:
    description: Paste the access token. The API accepts "Bearer " followed by the
      access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"